requirements.txt
atlassian-python-api>=4.0.0
cryptography>=42.0.0
python-dotenv>=1.0.0
pandas>=2.2.0
jinja2>=3.1.0
requests>=2.28.0
